+++ src/main/resources/templates/generators/tagReport.vm
@@ -18,7 +18,9 @@
 <div class="container-fluid" id="report">
   <div class="row">
     <div class="col-md-10 col-md-offset-1">
-      #includeElements($tag.getElements(), true)
+      <div class="elements">
+        #includeElements($tag.getElements(), true)
+      </div>
     </div>
   </div>
 </div>
