--- src/test/resources/net/masterthought/cucumber/no_step_results.json
@@ -1,85 +1,0 @@
-[
-    {
-        "id": "psa-product-testing",
-        "description": "As a PND Customer\nI want to make traffic information service request to Perseus application\nSo that i can receive the real time traffic, weather, safety camera services",
-        "name": "PSA product testing",
-        "keyword": "Feature",
-        "line": 1,
-        "elements": [
-            {
-                "id": "psa-product-testing;initsession-and-getmessage-request",
-                "description": "",
-                "name": "InitSession and GetMessage Request",
-                "keyword": "Scenario",
-                "line": 6,
-                "steps": [
-                    {
-                        "name": "a \"PSA\" PSA customer",
-                        "keyword": "Given ",
-                        "line": 7
-                    },
-                    {
-                        "name": "PSA customer make the InitSession request to the Perseus",
-                        "keyword": "When ",
-                        "line": 8
-                    },
-                    {
-                        "name": "the Perseus must respond with the response message and return code 201 to PSA",
-                        "keyword": "Then ",
-                        "line": 9
-                    },
-                    {
-                        "name": "the initSessionRequest is set to \"true\" and TTS is set to \"true\"",
-                        "keyword": "And ",
-                        "line": 10
-                    },
-                    {
-                        "name": "the tpeg response is set to \"true\" and the message format is \"TPEGBinary\" for PSA",
-                        "keyword": "And ",
-                        "line": 11
-                    },
-                    {
-                        "name": "the expected App is \"SAA\", \"WEA\", \"TEC\" in PSA",
-                        "keyword": "And ",
-                        "line": 12
-                    },
-                    {
-                        "name": "PSA customer make the GetMessage request to the Perseus",
-                        "keyword": "When ",
-                        "line": 13
-                    },
-                    {
-                        "name": "the Perseus must respond with the response message and return code 200 to PSA",
-                        "keyword": "Then ",
-                        "line": 14
-                    },
-                    {
-                        "name": "the get message content length is greater than 100 in PSA",
-                        "keyword": "And ",
-                        "line": 15
-                    },
-                    {
-                        "result": {
-                            "duration": 280197792,
-                            "status": "passed"
-                        },
-                        "name": "the number of tec,saa,wea messages should be greater than 0 in PSA",
-                        "keyword": "And ",
-                        "line": 16,
-                        "match": {
-                            "arguments": [
-                                {
-                                    "val": "0",
-                                    "offset": 58
-                                }
-                            ],
-                            "location": "PSATestStepDefs.the_number_of_tec_saa_wea_messages_should_be_greater_than_in_PSA(int)"
-                        }
-                    }
-                ],
-                "type": "scenario"
-            }
-        ],
-        "uri": "com/tomtom/perseus/uat/cucumber/psa.feature"
-    }
-]
