--- src/test/resources/net/masterthought/cucumber/with_no_step_results.json
@@ -1,85 +1,0 @@
-[
-    {
-        "id": "some-random-file",
-        "description": "As some kind of thing I want to do something so that I can do something",
-        "name": "Some product testing",
-        "keyword": "Feature",
-        "line": 1,
-        "elements": [
-            {
-                "id": "some-product-testing;do-request",
-                "description": "",
-                "name": "Do Request",
-                "keyword": "Scenario",
-                "line": 6,
-                "steps": [
-                    {
-                        "name": "a \"Big\" customer",
-                        "keyword": "Given ",
-                        "line": 7
-                    },
-                    {
-                        "name": "Big customer make the request to the Elephant",
-                        "keyword": "When ",
-                        "line": 8
-                    },
-                    {
-                        "name": "the Elephant must respond with the response message and return code woops",
-                        "keyword": "Then ",
-                        "line": 9
-                    },
-                    {
-                        "name": "the Request is set to \"true\" and Cats is set to \"true\"",
-                        "keyword": "And ",
-                        "line": 10
-                    },
-                    {
-                        "name": "the super response is set to \"true\" and the message format is \"Cars\" for Dogs",
-                        "keyword": "And ",
-                        "line": 11
-                    },
-                    {
-                        "name": "the expected Name is \"Wang\", \"Ping\", \"Bong\" in Elephant",
-                        "keyword": "And ",
-                        "line": 12
-                    },
-                    {
-                        "name": "Elephant customer make the Aaarg request to the Elephant",
-                        "keyword": "When ",
-                        "line": 13
-                    },
-                    {
-                        "name": "the Dog must respond with the response message and return code ouch to Mummy",
-                        "keyword": "Then ",
-                        "line": 14
-                    },
-                    {
-                        "name": "the get message content length is greater than Dogs in Sheep",
-                        "keyword": "And ",
-                        "line": 15
-                    },
-                    {
-                        "result": {
-                            "duration": 280197792,
-                            "status": "passed"
-                        },
-                        "name": "the number of dogs,cats,pea messages should be greater than 99 in Dogs",
-                        "keyword": "And ",
-                        "line": 16,
-                        "match": {
-                            "arguments": [
-                                {
-                                    "val": "0",
-                                    "offset": 58
-                                }
-                            ],
-                            "location": "ElephantSteps.the_number_of_dogs_cats_pea_messages_should_be_greater_than_in_Dogs(int)"
-                        }
-                    }
-                ],
-                "type": "scenario"
-            }
-        ],
-        "uri": "com/elephants/dogs/uat/cucumber/elephants.feature"
-    }
-]
